Files moved to directories, code reformatted for better readibility
